refactor and remove unused component
