+++ src/wallet/index.ts
@@ -2,13 +2,11 @@
 export { ConnectWallet } from './components/ConnectWallet';
 export { Wallet } from './components/Wallet';
 export { WalletDropdown } from './components/WalletDropdown';
-export { WalletMenu } from './components/WalletMenu';
 export { WalletBottomSheet } from './components/WalletBottomSheet';
 export { WalletDropdownBaseName } from './components/WalletDropdownBaseName';
 export { WalletDropdownDisconnect } from './components/WalletDropdownDisconnect';
 export { WalletDropdownFundLink } from './components/WalletDropdownFundLink';
 export { WalletDropdownLink } from './components/WalletDropdownLink';
-export { WalletWithMobileDrawer } from './components/WalletWithMobileDrawer';
 export { isValidAAEntrypoint } from './utils/isValidAAEntrypoint';
 export { isWalletACoinbaseSmartWallet } from './utils/isWalletACoinbaseSmartWallet';
 export type {
